fix send adventure fail (#176)

+++ MainCore/Tasks/Update/UpdateInfo.cs
@@ -151,10 +151,13 @@ namespace MainCore.Tasks.Update
             var adventures = context.Adventures.Count(x => x.AccountId == AccountId);
             if (numberAdventure == 0)
             {
-                var heroAdventures = context.Adventures.Where(x => x.AccountId == AccountId).ToList();
+                if (adventures != 0)
+                {
+                    var heroAdventures = context.Adventures.Where(x => x.AccountId == AccountId).ToList();
 
-                context.Adventures.RemoveRange(heroAdventures);
-                context.SaveChanges();
+                    context.Adventures.RemoveRange(heroAdventures);
+                    context.SaveChanges();
+                }
             }
             else if (adventures != numberAdventure)
             {
